amd/requirements.txt
torch-directml
torchvision
pillow
tqdm